Updates to BaseViewModels and Sample
Updates to:
* Allow for EntityData and EntityDataAlwaysLatest to be used. Currently
it was forcing EntityData in BaseTableDataStore. EntityData now
inherits from EntityDataAlwaysLatest to support this. BaseTable will
deal with EntityDataAlwaysLatest, so now due to the power of
polymorphism it will deal with EntityData also
* Added/removed items to the collection when using the
AddItemAsync/DeleteItemAsync so the collection was in sync with updates
* Added Commands to the BaseViewModel so there is less work to do.
Added Save/Delete/AddNew commands
* Fixed the SelectedItem mode in the ToDosViewModel with two-way
bindings
* Changes the ViewModel’s AddNewItem/UpdateItemAsync/DeleteItemAsync to
be virtual so Navigation hooks can be added in.
* Added a simple navigation system. It should be extended for more
complex situations
* Optimised the ViewCell binding to use Grid instead of StackLayouts
and added binding